split each day's stuct into it's own file
