+++ aoc-2020/Day02/Day02.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Linq;
+using System.IO;
+using System.Text.RegularExpressions;
+
+namespace aoc2020
+{
+	struct Day02 : IAOCProgram
+	{
+		public void Run()
+		{
+			Regex expression = new Regex(@"(?<min>\d+)-(?<max>\d+)\s(?<char>\w):\s(?<password>\w+)");
+			var values02 = File.ReadAllLines(@"Day02/input.txt")
+				.Select(line => {
+					var match = expression.Match(line);
+					if (match.Success) {
+						return new Passcode(
+							int.Parse(match.Groups["min"].Value),
+							int.Parse(match.Groups["max"].Value),
+							match.Groups["char"].Value[0],
+							match.Groups["password"].Value
+						);
+					}
+
+					Console.WriteLine($"Failed to parse: {line}");
+					return new Passcode();
+				});
+
+			var sledChecker = new SLEDPasscodeChecker();
+			Console.WriteLine($"pt1 Number of valid passcodes: {values02.Where(passcode => sledChecker.IsValid(passcode)).Count()}");
+			var otcChecker = new OTCPasscodeChecker();
+			Console.WriteLine($"pt2 Number of valid passcodes: {values02.Where(passcode => otcChecker.IsValid(passcode)).Count()}");
+		}
+	}
+}
