+++ aoc-2020/Day03/Day03.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.IO;
+
+namespace aoc2020
+{
+	struct Day03 : IAOCProgram
+	{
+		public void Run()
+		{
+			var map = File.ReadAllLines(@"Day03/input.txt");
+
+			var toboggan = new Toboggan();
+			long trees = toboggan.Sled(map, new Slope(3, 1));
+			Console.WriteLine($"pt1 Tree Count: {trees}");
+
+			trees = 1;
+			trees *= toboggan.Sled(map, new Slope(1, 1));
+			trees *= toboggan.Sled(map, new Slope(3, 1));
+			trees *= toboggan.Sled(map, new Slope(5, 1));
+			trees *= toboggan.Sled(map, new Slope(7, 1));
+			trees *= toboggan.Sled(map, new Slope(1, 2));
+			Console.WriteLine($"pt2 Tree Count: {trees}");
+		}
+	}
+}
