+++ aoc-2020/IAOCProgram.cs
@@ -1,0 +1,7 @@
+﻿namespace aoc2020
+{
+	interface IAOCProgram
+	{
+		void Run();
+	}
+}
