+++ aoc-2020/aoc-2020.csproj
@@ -32,17 +32,24 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Day01\Day01.cs" />
     <Compile Include="Day01\SumSearcher.cs" />
+    <Compile Include="Day02\Day02.cs" />
     <Compile Include="Day02\Passcode.cs" />
+    <Compile Include="Day03\Day03.cs" />
     <Compile Include="Day03\Slope.cs" />
     <Compile Include="Day03\Toboggan.cs" />
+    <Compile Include="Day04\Day04.cs" />
     <Compile Include="Day04\PassportInfo.cs" />
     <Compile Include="Day04\IPassportVarifier.cs" />
     <Compile Include="Day04\BasicVarifier.cs" />
     <Compile Include="Day04\FullVarifier.cs" />
+    <Compile Include="Day05\Day05.cs" />
     <Compile Include="Day05\Ticket.cs" />
     <Compile Include="Day05\TicketConversion.cs" />
+    <Compile Include="Day06\Day06.cs" />
     <Compile Include="Day06\CultureDeclaration.cs" />
+    <Compile Include="IAOCProgram.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Day01\input.txt">
